fix: make GetRawRoot() work properly with GitHub
GetRawRoot() failed to return the correct root when passed GitHub
URLs ending with .git (eg. https://github.com/foo/bar.git), which
are supposed to be equivalent to the same URL without ".git".

That was disguised by the fact that GetRepo() erroneously stripped
the ".git" suffix from the passed URL and made it work when called
before GetRawRoot() with the same argument.